Create Flowers.java
The flower shop near his house sells flowers of N types. It is found that the store has Ai flowers of the i-th type. We like odd numbers. Therefore, we have decided that there should be an odd number of flowers of each type in the bouquet, and the total number of flowers in the bouquet should also be odd.
Determine the maximum number of flowers the bouquet can consist of.
Input
The first line contains an integer N — the number of types of flowers that are sold in the store
The second line contains N integers— the number of flowers of each type

1 <= N <= 100000
1 <= Ai <= 1000
Output
Print one number — the maximum number of flowers the bouquet can consist of.