Create Sum Sub Array.java
You are given an array Arr of N integers. A subarray is good if the sum of elements of that subarray is greater than or equal to K. Print the length of good subarray of minimum length.
Input
First line of input contains N and K.
Second line of input contains N integers representing the elements of the array Arr.

Constraints
1 <= N <= 100000
1 <= Arr[i] <= 100000
1 <= K <= 1000000000000
Output
Print the length of good subarray of minimum length.

Sample input
5 12
2 3 2 5 5 

Sample output
3

Explanation : 
Subarray from index 3 to 5 has sum 12 and is therefore good and its length(3) is minimum among all possible good subarray.